pyproject.toml
[project]
name = "multi-platform-device-api"
version = "0.1.0"
description = "Django API for multi-platform device management"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "Django>=5.1",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.4",
    "django-cors-headers>=4.4",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.5.0",
    "pre-commit>=3.7.0",
    "coverage>=7.5.0",
]

[tool.ruff]
line-length = 88
target-version = "py312" # Ruff might not explicitly support py314 keyword yet, keeping safe recent version

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["config", "api"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
